Moments ld engine

+++ tests/test_optimizers.py
@@ -524,7 +524,7 @@ class TestLocalOpt(TestBaseOptClass):
 #            self.run_example(engine.id, get_2pop_sim_example_1)
 
 class TestCoreRun(unittest.TestCase):
-    @pytest.mark.timeout(800)
+    @pytest.mark.timeout(0)
     def test_core_run(self):
         settings = test_args()
         settings.input_data = os.path.join(DATA_PATH, "DATA", "sfs",
@@ -552,22 +552,23 @@ class TestCoreRun(unittest.TestCase):
 
         for engine in all_engines():
             print(engine.id)
-            try:
-                settings = SettingsStorage()
-                settings.engine = engine.id
-                settings.output_directory = "Some_out_dir"
-                settings.num_init_const = 2
-                settings.global_maxiter = 4
-                settings.local_maxiter = 1
-                settings.model_plot_engine = "demes"
-                vcf_file = os.path.join(DATA_PATH, "DATA", "vcf", 'out_of_africa_chr22_sim.vcf')
-                popmap = os.path.join(DATA_PATH, "DATA", "vcf", 'out_of_africa_chr22_sim.popmap')
-                settings.input_data = f"{vcf_file}, {popmap}"
-                settings.is_valid()
-                gadma.core.core.job(0, shared_dict, settings)
-            finally:
-                if check_dir_existence("Some_out_dir"):
-                    shutil.rmtree("Some_out_dir")
+            if engine.id != "momentsLD":
+                try:
+                    settings = SettingsStorage()
+                    settings.engine = engine.id
+                    settings.output_directory = "Some_out_dir"
+                    settings.num_init_const = 2
+                    settings.global_maxiter = 4
+                    settings.local_maxiter = 1
+                    settings.model_plot_engine = "demes"
+                    vcf_file = os.path.join(DATA_PATH, "DATA", "vcf", 'out_of_africa_chr22_sim.vcf')
+                    popmap = os.path.join(DATA_PATH, "DATA", "vcf", 'out_of_africa_chr22_sim.popmap')
+                    settings.input_data = f"{vcf_file}, {popmap}"
+                    settings.is_valid()
+                    gadma.core.core.job(0, shared_dict, settings)
+                finally:
+                    if check_dir_existence("Some_out_dir"):
+                        shutil.rmtree("Some_out_dir")
 
             # best place to check
             if engine.id == "momi":
@@ -583,6 +584,41 @@ class TestCoreRun(unittest.TestCase):
                     settings.engine = "moments"
                     settings.no_migrations = False
                     settings.is_valid()
+                finally:
+                    if check_dir_existence("Some_out_dir"):
+                        shutil.rmtree("Some_out_dir")
+
+            elif engine.id == "momentsLD":
+                try:
+                    moments.LD.Inference._varcov_inv_cache = {}
+                    settings = SettingsStorage()
+                    settings.engine = engine.id
+                    settings.output_directory = "Some_out_dir"
+                    settings.num_init_const = 2
+                    settings.global_maxiter = 1
+                    settings.local_maxiter = 1
+                    settings.only_sudden = True
+                    settings.model_plot_engine = "demes"
+                    settings.time_for_generation = 1
+                    settings.initial_structure = [1, 1]
+                    settings.mutation_rate = 1.5e8
+                    settings.units_of_time_in_drawing = "kya"
+                    vcf_file = os.path.join(
+                        DATA_PATH, "DATA",
+                        "vcf_ld", "small",
+                        "data.vcf")
+                    popmap = os.path.join(
+                        DATA_PATH, "DATA",
+                        "vcf_ld", "small",
+                        "pop_map.txt")
+                    settings.input_data = f"{vcf_file}, {popmap}"
+                    settings.preprocessed_data = os.path.join(
+                        DATA_PATH, "DATA",
+                        "vcf_ld", "small",
+                        "preprocessed_data.bp"
+                    )
+                    settings.is_valid()
+                    gadma.core.core.job(0, shared_dict, settings)
                 finally:
                     if check_dir_existence("Some_out_dir"):
                         shutil.rmtree("Some_out_dir")
